Rename HearingDetailsTable to HearingTable
