--- src/features/CourtCaseDetails/Tabs/Panels/HearingDetailsTable.tsx
@@ -1,68 +1,0 @@
-import { Hearing } from "@moj-bichard7-developers/bichard7-next-core/build/src/types/AnnotatedHearingOutcome"
-
-import { Table } from "govuk-react"
-import courtTypes from "@moj-bichard7-developers/bichard7-next-data/dist/data/court-type.json"
-import { getDisplayDateFromDateOrString } from "utils/formattedDate"
-
-interface HearingDetailsTableProps {
-  hearing: Hearing
-}
-
-interface HearingDetailsTableRowProps {
-  header: string
-  value: string | null | undefined
-}
-
-const HearingDetailsTableRow = ({ header, value }: HearingDetailsTableRowProps) => (
-  <Table.Row>
-    <Table.Cell>
-      <b>{header}</b>
-    </Table.Cell>
-    <Table.Cell>{value}</Table.Cell>
-  </Table.Row>
-)
-
-export const HearingDetailsTable = ({ hearing }: HearingDetailsTableProps) => {
-  const getLanguage = (language: string) => {
-    switch (language) {
-      case "E":
-        return `English(${language})`
-      case "W":
-        return `Welsh (${language})`
-      case "D":
-        return `Don't know (${language})`
-      default:
-        return language
-    }
-  }
-
-  const getCourtType = (courtCode: string | null | undefined) => {
-    let courtTypeWithDescription = courtCode
-    courtTypes.forEach((type) => {
-      if (type.cjsCode === courtCode) {
-        courtTypeWithDescription = `${courtCode} (${type.description})`
-      }
-    })
-    return courtTypeWithDescription
-  }
-
-  return (
-    <Table>
-      <HearingDetailsTableRow header="Court location" value={hearing.CourtHearingLocation.OrganisationUnitCode} />
-      <HearingDetailsTableRow header="Date of hearing" value={getDisplayDateFromDateOrString(hearing.DateOfHearing)} />
-      <HearingDetailsTableRow header="Time of hearing" value={hearing.TimeOfHearing} />
-      <HearingDetailsTableRow header="Defendant present" value={hearing.DefendantPresentAtHearing} />
-      <HearingDetailsTableRow header="Source reference document name" value={hearing.SourceReference.DocumentName} />
-      <HearingDetailsTableRow header="Source reference identification" value={hearing.SourceReference.UniqueID} />
-      <HearingDetailsTableRow header="Source reference document type" value={hearing.SourceReference.DocumentType} />
-      <HearingDetailsTableRow header="Court type" value={getCourtType(hearing.CourtType)} />
-      <HearingDetailsTableRow header="LJA code" value={hearing.CourtHouseCode.toString()} />
-      <HearingDetailsTableRow header="Court name" value={hearing.CourtHouseName} />
-      <HearingDetailsTableRow header="Hearing language" value={getLanguage(hearing.HearingLanguage)} />
-      <HearingDetailsTableRow
-        header="Documentation language"
-        value={getLanguage(hearing.HearingDocumentationLanguage)}
-      />
-    </Table>
-  )
-}
